Allow filterbank_header to write headers when the machine is FAKE
There seems to have been a check for whether or not the machine is an actual machine, but ignoring the fact that not every machine has an alias. As a result, when the machine is set to FAKE, filterbank_header was silently refusing to write a header to disk. This changes the check to be such that the machine ID is greater than 0, the only place where I can find that this is not the case is in the preparation of reading psrfits files, at readpsrfits_hd.c#32, which has been changed to 0 as a result.